Move mocks for course unit tests to tests/samples/
Move the Python file of mocks for unit tests for the course module to
the tests/samples directory so that it's in the same directory as most
other mocks. This is part of a larger effort to migrate all mocks in
tests/mocks to tests/samples so that all mocks are under the same
directory.
